requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
colorama==0.4.5
distlib==0.3.5
Django==4.0.6
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.7.1
iniconfig==1.1.1
model-bakery==1.6.0
packaging==21.3
pipenv==2022.7.4
pkg_resources==0.0.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-watch==4.2.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
virtualenv==20.15.1
virtualenv-clone==0.5.7
watchdog==2.1.9
